feat: Implement network packet handling and parsing

+++ src/tools/tcp.rs
@@ -1,5 +1,5 @@
 #[allow(dead_code)]
-pub const TCP_PROTOCOL_ID: u16 = 0x06;
+pub const TCP_PROTOCOL_ID: u8 = 0x06;
 
 pub struct TcpHeader {
     pub src_port: u16,
